Виправлено помилки. З радіо датчика передаеться температура повністю.

+++ NRF24l01P_Remote/ds18b20.h
@@ -9,7 +9,7 @@
 const char fract[] = {0,1,1,2,2,3,4,4,5,6,6,7,7,8,9,9}; // дискретність, знак після коми
                                                         //  0/16..., 1/16..., 2/16......
 
-uint8_t readTemp_Single(uint8_t *buf, uint8_t *minus);
+uint8_t readTemp_Single(uint16_t *buf, uint8_t *minus);
 void init_ds18b20(void);
 
 #endif	/* XC_HEADER_TEMPLATE_H */
